ATM Program
This code is a program for a ATM banking system that allows the user to use their card and PIN to access their account. The program has a menu of options, including balance inquiry, withdrawal, deposit, and money transfer. If the user selects balance inquiry, the program displays the total balance in the account. If the user selects money transfer, they must enter the account number and amount to be transferred. The transfer will be made if the account number is verified and the amount is a multiple of 100 and less than 50000. If the transfer amount is not in multiples of 100, a message will display stating that it is not in the multiples of 100. If the user selects withdrawal, they must enter the amount, which must be in multiples of 100. If the balance is sufficient, the program will display a message to collect the amount and update the balance. If the balance is insufficient, the program will display a message stating that the amount is not available. If the user selects deposit, they must enter the amount, and the program will display a message stating that the deposit has been made and update the balance.